+++ src/mapping-entry.ts
@@ -14,6 +14,7 @@ export interface Paths {
  * The array is sorted by longest prefix.
  * Having an array with entries allows us to keep a sorting order rather than
  * sort by keys each time we use the mappings.
+ *
  * @param absoluteBaseUrl
  * @param paths
  * @param addMatchAll
@@ -32,7 +33,7 @@ export function getAbsoluteMappingEntries(
     absolutePaths.push({
       pattern: key,
       paths: paths[key].map((pathToResolve) =>
-        path.join(absoluteBaseUrl, pathToResolve)
+        path.resolve(absoluteBaseUrl, pathToResolve)
       ),
     });
   }