+++ CHANGELOG.md
@@ -6,7 +6,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
-- Ability to use absolute paths. (#184)
+- Ability to use absolute paths. [#184](https://github.com/dividab/tsconfig-paths/pull/184)
 
 ## [3.11.0] - 2021-08-24